# Improve CompareOrdinalIgnoreCaseHelper performance for matching chars