@jkotas 2017-01-24
@bbowyersmyth Would you mind porting this to CoreRT as well?