@bbowyersmyth 2017-01-25
Thanks @jkotas. The CoreRT version is a bit different but I'll look into it